Remove dependency injection for plugins and validate within Character
The previous method was to assume that a plugin module would export a function, to which the caller could pass the Plugin base class.

Now a plugin module will be assumed to export a complete class. The plugin module itself can access the base class with `character/plugin`

Plugin API validation is also moved into the Character core module to reduce interference with plugin API.